remove mode handling as view block arg; handle with with_nest_mode block.
